moved PostConstruct annotation processing to BeanPostProccesor
- BeanPostProcessor implementation - PostConstructBeanPostProcessor for processing the PostConstruct annotation has been created; Processing logic removed from GenericApplicationContext;
- added tests for a new class;
- change pom.xml :
	+ Removed unnecessary dependency on Mokito;
	+ upgraded versions of junit and slf4j;
