+++ pom.xml
@@ -28,20 +28,14 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.13.1</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-all</artifactId>
-            <version>1.9.5</version>
+            <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
         <!-- Logger -->
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-simple</artifactId>
-            <version>1.8.0-beta1</version>
+            <version>2.0.3</version>
         </dependency>
         <!-- Lombok -->
         <dependency>
